requirements.txt
azure-cosmos
azure-identity
flask
gunicorn
